% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{performance}
\alias{performance}
\title{Estimate validation performance}
\usage{
performance(
  object,
  method = c("naive", "gold_only", "OR", "AIPW"),
  metrics = c("auc", "brier", "cal_intercept", "cal_slope"),
  outcome_formula = gold_outcome ~ proxy_outcome * qlogis_pred,
  verification_formula = verified ~ proxy_outcome + qlogis_pred
)
}
\arguments{
\item{object}{A \code{goldval_object}.}

\item{method}{Character vector of methods. Supported methods are \code{"naive"},
\code{"gold_only"}, \code{"OR"}, and \code{"AIPW"}.}

\item{metrics}{Character vector of metrics. Supported metrics are \code{"auc"},
\code{"brier"}, \code{"cal_intercept"}, and \code{"cal_slope"}.}

\item{outcome_formula}{Formula for the gold-outcome regression model used by
corrected methods.}

\item{verification_formula}{Formula for the verification model used by AIPW.}
}
\value{
A \code{goldval_performance} object.
}
\description{
Estimate validation performance
}
