% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{calibration}
\alias{calibration}
\title{Estimate calibration summaries and flexible calibration curves}
\usage{
calibration(
  object,
  method = c("naive", "gold_only", "OR", "AIPW"),
  type = c("weak", "curve"),
  outcome_formula = gold_outcome ~ proxy_outcome * qlogis_pred,
  verification_formula = verified ~ proxy_outcome + qlogis_pred,
  curve_df = 4,
  grid = NULL
)
}
\arguments{
\item{object}{A \code{goldval_object}.}

\item{method}{Character vector of methods. Supported methods are \code{"naive"},
\code{"gold_only"}, \code{"OR"}, and \code{"AIPW"}.}

\item{type}{Calibration outputs to return: \code{"weak"} for logistic
recalibration intercept/slope and/or \code{"curve"} for a flexible calibration
curve.}

\item{outcome_formula}{Formula for the gold-outcome regression model.}

\item{verification_formula}{Formula for the verification model.}

\item{curve_df}{Degrees of freedom for the natural spline calibration curve.}

\item{grid}{Optional prediction-risk grid.}
}
\value{
A \code{goldval_calibration} object.
}
\description{
Estimate calibration summaries and flexible calibration curves
}
